requirements.txt
flask
flask_cors
line-bot-sdk
python-dotenv
google-cloud-logging
google-cloud-firestore
google-cloud-storage
gunicorn
pyimgur
opencv-contrib-python
Pillow
torchgpipe
matplotlib
requests
fvcore
fire